refactor : WinnerLotto Bonus 의 책임 WinnerLotto 로 위임
