+++ src/test/java/lotto/domain/WinnerLottoTest.java
@@ -1,9 +1,11 @@
 package lotto.domain;
 
+import static lotto.domain.WinnerLotto.BONUS_NUMBER_OVERLAPPING_EXCEPTION;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.HashSet;
 import java.util.Set;
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -17,7 +19,7 @@ class WinnerLottoTest {
         // given
         Lotto lotto = new Lotto(1, 2, 3, 4, 5, 6);
         int bonusNumber = 7;
-        WinnerLotto winnerLotto = new WinnerLotto(lotto, bonusNumber);
+        WinnerLotto winnerLotto = new WinnerLotto(lotto, LottoNumber.valueOf(bonusNumber));
 
         // when
         int result = winnerLotto.countOfMatch(createLotto());
@@ -43,12 +45,24 @@ class WinnerLottoTest {
     void contains_bonus(int bonusNumber, boolean expected) {
         // given
         Lotto lotto = new Lotto(1, 2, 3, 4, 5, 6);
-        WinnerLotto winnerLotto = new WinnerLotto(lotto, bonusNumber);
+        WinnerLotto winnerLotto = new WinnerLotto(lotto, LottoNumber.valueOf(bonusNumber));
 
         // when
         boolean result = winnerLotto.containsBonus(createLotto());
 
         // then
         assertThat(result).isEqualTo(expected);
+    }
+
+    @Test
+    @DisplayName("당첨 로또와 보너스 숫자가 겹치면 예외를 던진다.")
+    void create_bonus_exception() {
+        // given
+        Lotto lotto = new Lotto(1, 2, 3, 4, 5, 6);
+
+        // when // then
+        Assertions.assertThatThrownBy(() -> new WinnerLotto(lotto, LottoNumber.valueOf(6)))
+                .isExactlyInstanceOf(IllegalArgumentException.class)
+                .hasMessage(BONUS_NUMBER_OVERLAPPING_EXCEPTION);
     }
 }
